Update introduce-class.md
function mapStateToProps(state) {
  return { todos: state.${namespace}.todos };
}
connect(mapStateToProps)(App);
---------------------------------------------------------------------------------------------------
I really confused that I follow you to code this sample ,but it dosen't work. Because   the param of  the function `mapStateToProps`  is `state` and  get the state number value throught it must add the namespace  , or it would take nothing but undefined.

Thanks!